Revert "Fix panels starting below other windows."
This reverts commit d6f56253b2d2ffd8be1f6283fac2a3e13a45130c.

This didn't work as expected, as windows' input was still above panels
even though windows looked like they were behind panels, whenever the
problem happened. When fullscreen mode is initiated in evince, this
commit also caused windows to have their input in front
of evince's fullscreen window even though the fullscreen window appears
to be in front of them. This in turn triggered a freeze when the user
clicks on the screen in evince while in fullscreen mode.