Cleanup linkage handling and use rpath for plugins.
rpath is generally bad because it results in stupid linkage which can
break with distribution upgrades (which is the reason for previous
commits which use ldconfig LIBDIR). However, it doesn't appear to be
against any distribution's policy to use rpath for "internal" libraries
such as plugins linking against each other - so instead of adding those
rather generic names to the ld.so cache, just specify an rpath, since there
is no way there would be multiple versions of a plugin at once (and even then
plugins would have to manually link).

This *does* stretch the boundaries allowed by distributions slightly however,
since other plugins usually have to link with libcomposite and libopengl
in order to load them, however we have our own system of safeguard checks
to ensure that the right plugins are loaded in any case (since the linkage
will change dynamically on dlopen () - the initial linkage is only there
so that dlopen () will not complain when it can't find the composite and
opengl libraries, but in that case, it wouldn't even matter anyways since
those plugins have to be loaded in order to be useful)

For distribution policy, see:
http://wiki.debian.org/RpathIssue
http://fedoraproject.org/wiki/Packaging/Guidelines#Beware_of_Rpath