Check if header token matches part of another word.
The prepending function would only check for the existence of part of the header op in the function string, and did not check if the matched string was actually part of another word. This caused fragment programs like this to fail:

program name: temporary
!!ARBfp1.0
TEMP temp, foo, bar;
FOO temp, bar, temp;

Which would be outputted like this:

TEMP temporary_temp, temporary_foo, temporary_bar;
FOO temporary_temporary_temp, bar, temporary_temporary_bar;

(since "temp" was matched in "temporary" incorreclty).

The new code now checks for a ".", ";" or "," or if the word is the last in the string before prepending it.

This bug was most noticable in the colorfilter plugin, where one of the filters could be named "negative" with a temporary variable called "neg"