Fix windows jumping around when clicking on parts that trigger move
Since we no longer have an unconditional passive grab on Button1
for active and raised windows, we need to check if a client
sent us a _NET_WM_MOVERESIZE ClientMessage and update the pointer
buffer on that.

Also clean up the pointer buffer handling code again - reduce
the number of roundtrips required. Remove debug message.