Add additional checks for features before running
We used to ship a shell script that checked for various things before
running compiz. However, this duplicated several checks that compiz
runs on its own and was slow due to being a shell script.

This patch adds in default options and plugins if none are provided on
the command line so running `compiz` continues to work like expected.